feat: create MListProducts
